+++ src/main/java/net/sf/jabref/gui/BasePanel.java
@@ -48,6 +48,7 @@ import net.sf.jabref.gui.actions.Actions;
 import net.sf.jabref.gui.actions.BaseAction;
 import net.sf.jabref.gui.actions.CleanupAction;
 import net.sf.jabref.gui.actions.CopyBibTeXKeyAndLinkAction;
+import net.sf.jabref.gui.actions.CopyPreviewOfSelectedEntriesToClipboardAction;
 import net.sf.jabref.gui.bibtexkeypattern.SearchFixDuplicateLabels;
 import net.sf.jabref.gui.desktop.JabRefDesktop;
 import net.sf.jabref.gui.entryeditor.EntryEditor;
@@ -494,6 +495,8 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
 
         // The action for copying the BibTeX keys as hyperlinks to the urls of the selected entries
         actions.put(Actions.COPY_KEY_AND_LINK, new CopyBibTeXKeyAndLinkAction(mainTable));
+
+        actions.put(Actions.COPY_PREVIEW_OF_SELECTED_ENTRIES_TO_CLIPBOARD, new CopyPreviewOfSelectedEntriesToClipboardAction(frame));
 
         actions.put(Actions.MERGE_DATABASE, new AppendDatabaseAction(frame, this));